implement id
This is a large change, because it required adding mutability.

"id" only makes sense if you have data that can mutate, such that
changes through one reference are visible through another reference. But
before this commit, the data model we were using was immutable. That
immutability was a natural result of implementing in Haskell, where all
values are immutable.

To model mutable data, I used IORef. Pairs are now indirected through an
IORef, which means that they can only be read and written to inside the
IO monad. This required plumbing monads and references all over the
code.

To mitigate the impact of plumbing IORefs everywhere, this commit
introduces the concept of a MonadRef: a monad which allows reading and
writing references. Even though in practice this monad is sometimes IO,
the helper code in Data.hs can be written in terms of an interface that
is much smaller than IO: an interface of reading and writing references
in a monad.

In Parse.hs, I used Identity as the MonadRef. It seemed overly
complicated to plumb the IO monad around through the parsing code.
Instead of doing that, we plumb Identity around (which is easy), and
then batch convert all of the Identitys to new IORefs at the end of
parsing.

One of the tricky bits was figuring out what to do with EvalState.
EvalState contains Objects, which now contain IORefs, which means it can
only be manipulated in IO. On earlier passes, this was even stronger: it
could only be manipulated in EvalMonad. Some of my early attempts put
monadic values inside EvalState, because I did not know what else to do.
But if there are monadic values inside the state, then the computation
to get a piece of the state is `m(m a)`, where the inner `m` is not a
computation I actually wanted or cared about. I thought I was stuck with
this, because I needed to get an initial state in somehow and the only
way I had written initial states in the past was by passing them in to
`runState`. Values passed in to `runState` have to be created outside
the monad, but I can't create IORef outside at least the IO monad (and
initially, I couldn't create Pairs outside the EvalMonad). But then
@rjmk pointed out that:

  1. I could pass in an empty state to runState, and I _can_ create the
  empty state outside of IO.
  2. It should be possible to create objects just in IO, rather than the
  full EvalMonad.

These were very helpful observations, which together let me remove
monadic values from my state.

There is more I could say about this change, but I am running out of
steam on this message. Perhaps I will write more in another format.