start interpreting bel.bel 🎉
Baby steps. This adds a test for the first definition of bel.bel, and
the test is failing. But it's close.

This commit adds test/part-of-bel.bel, which is a small part of Paul
Graham's bel.bel. It also adds a test that interprets
test/part-of-bel.bel and runs tests against the interpretation.

This commit also adds comment parsing.

This commit also partially implements "def".