fix function argument evaluation
This is a benefit of starting to interpret the prelude. I immediately
uncovered a way in which I had misunderstood the spec.

When arguments are passed to functions, they are evaluated first. But
the function might also evaluate its arguments internally (eg by passing
them to another function). Evaluation is not idempotent. For example, ''a
evaluates to 'a. So if we pass ''a to a function, it gets evaluated to 'a,
and then if it gets passed on to an internal function, it gets evaluated
again to a, meaning the inner function sees a different value than the
outer function. This behavior seems weird, and is also, it turns out,
contradicted by this example from the spec:

    > (no 'a)
    nil

My guess now is that after we evaluate the arguments to a function, we
quote them before passing them in. This lets side-effects happen, but
doesn't change the value.

I could still be confused, but this is my working interpretation.