fix list evaluation
We weren't evaluating lists correctly in all cases. In particular, when
the head of the list was something that evaluated to a primitive, but
was not yet in literal form (eg '(lit prim join)), we would fail to
treat it as a primitive function call.

The relevant logic was spread out between long inline blocks of
implementations of primitives and special forms, which made it harder to
understand. I pulled the implementations of primitives and special forms
out of `evaluate`, which made the core pair evaluation logic easier to
follow. This change also resulted in a moderate increase in code size,
which is disappointing but expected for pulling out inlined code.

This change also adds more of bel.bel to the test file, and adds a bunch
of commented-out tests for it, plus one new live test.