implement destructuring of function arguments
The cons function from bel.bel requires destructuring, which I hadn't
implemented yet. Implementing it was sufficient to get the bel.bel tests
passing for cons.

I also added pretty printing for quotes, so (quote a) prints as 'a.

I also added some more tests that are commented out for now, but will be
enabled as I get further through the spec.