treat pairs with vmark in their car as variables
Previously, I was only allowing symbols to be variables. That is to say,
only symbols were allowed in the "var" slot of (var . val) pairs in
globe and scope, in closure environments, in parameter lists, and in
calls to dyn and set.

But, as The Bel Language guide says:

  Any pair with vmark in its car is treated by Bel as a variable.

It took me some time to understand what was going on with vmark (and its
companion, uvar). But now I do, and here is my implementation.

We can new interpret bel.bel up through the definition of or.