silently ignore attempts to overwrite vmark
vmark has a special meaning. It refers to an object that is built-in to
the interpreter. Overwriting it with anything else will cause (vmark)
and (uvar) not to do what they are supposed to.

bel.bel defines vmark by calling set. Since we already have a definition
of vmark built in, we will just ignore it.