create an interface for optimized functions
This lets me write Bel functions in Haskell, while still having them
appear to be regular Bel lists if they are ever inspected.

This commit doesn't include any instances of these functions. I had a
use for them in mind when I started, but now I have a different plan
that does not require these. I'm committing this code anyway, because it
seems like a useful tool that I will want in the future.