native implementations of cons and append
This is a performance improvement, but more importantly it gives me
implementations of cons and append that I can use even when I'm not
loading bel.bel. This in turn means I can implement backtick in terms of
them. That in turn will mean that I can implement backtick in a way that
will not cause problems when passed to the bel-in-bel interpreter.

The append algorithm is still a slow one, despite the constant-factor
speedup from being implemented natively.