add more tests from the spec
And rearrange some of the tests at the bottom. The tests at the bottom
are ones that I implemented out-of-order. The tests above them should be
a contiguous chunk of every evaluation example written in the spec,
starting from the top of the spec. Now that this contiguous chunk is
catching up with my out-of-order implementations, we can slot the tests
into their places.