replace native prnum with native prsimple
I'm using Bel to solve Advent of Code problems now. For day 1 part 1, I
need to print a seven-digit number. The implementation from bel.bel, as
evaluated by the current version of this interpreter, takes much longer
than I am willing to wait to print such a number.

I tried to accelerate this in the previous commit by native-izing number
and prnum, but this was not sufficient. The Bel implementation of
prsimple calls numr and numi on its input when the input is a number. If
that number is seven digits, this decompresses the optimized number,
creates a list of seven-digit length, and then passes it to the (now
optimized) prnum which just re-compresses the list back to a number.

That is very slow and wasteful, so to avoid it I am also native-izing
prsimple. Now that I've done that, I don't really need a native prnum
anymore, so I am erasing it in this commit. I could have not committed
it in the first place (I finished implementing prnum and prsimple before
committing either), but I wanted to have a record of the prnum
implementation just in case I ever want it again.