passing a filename in CLI now interprets the file
Previously it parsed and executed the file and then put you in a REPL
with the resulting global names. Inconsistently with other entry points,
that one alone would _not_ load the definitions from bel.bel.

The interpreter now _does_ load the definitions from bel.bel in that
mode. It no longer dumps you into a REPL, and instead just executes the
file and then exits. It would be useful eventually to explore user files
in the REPL, but that will happen another time.