create mechanism to cache global state to a string
It's pretty simple. Just convert the Haskell representation of the state
to a Bel object, and then serialize that object with repr. This can the
be deserialized with parse and a conversion back from Bel object to
Haskell representation.

I'm hoping to use this to cache the global state after executing
bel.bel, so we don't have to re-execute it on every interpreter startup.