fix shared pair printing
There was a bug in printing (a . b) when b was both a shared pair and a
proper list. In that case, b would print without parens around it --
correct if it is not a shared pair, but not correct if it is. This fix
is a bit of a hack, but it didn't feel worth it to make a bigger change
here yet.