start repl and interpreter in <0.1 sec
Almost all of the startup latency was in executing bel.bel. We now no
longer need to execute bel.bel on startup, because we have the results
of executing it cached in the interpreter binary.

Side note: I also aspire to execute programs the size of bel.bel in <0.1
seconds. This is tricky in the case of bel.bel in particular because it
redefines some of my optimized native functions to be slower user-level
Bel functions. But now it doesn't matter much how long it takes to
execute bel.bel in particular, because it's cached on build. It takes
about 1.5 seconds to execute it during the build process.