fix shared pair printing better
Like f89a5ab727593f61fbf3156e51562c58e63b4a0d, this is a fix for the
case of printing (a . b) where b is a proper list and a shared pair. We
want to print this as (a . #n=b), where b has parens around it.
Previously we were not printing the dot. Now we print the dot.