add tests
At this point I think I have pulled all of the easily-extractable
examples from bellanguage.txt and belexamples.txt into tests.

Some of the tests are commented out because they are too slow or don't
work yet, but they are at least present in the file. The remaining
implementation work is to uncomment them one by one and get them to
pass.