uncomment slow tests
Some of them no longer add noticeable time to the test run.

Some of them are now broken, and stay commented :/

Some of them are still slow, and now run when and only when the env var
CHIME_INCLUDE_SLOW_TESTS is set to anything. This var is now set in CI.

One of them was duplicated.