fast numeric fns now work on non-numeric inputs
I optimized numeric functions a while ago by writing native Haskell
implementations that, unlike the Bel versions, didn't have to treat
numbers as being unary-encoded.

One problem with this optimization was that it made those functions stop
working on anything that was not a number. This is particularly a
problem for eg <, which can be defined for many non-numeric inputs.

This commit changes the way the optimized functions are used, so that
when they receive an input outside of the domain they can handle, we
fall back to the Bel implementation from the spec.