table references are now set-able
I took a different approach from the reference interpreter here. The
reference interpreter has a concept of locfns, which are called when
applying a car, cdr, or tab expression.

In the case of this interpreter, setting car and cdr expressions was
already implemented via a different mechanism. It didn't quite make
sense to directly import the locfn idea just for tables. Instead, this
implementation makes use of the following insight:

    (cdr:tabloc tab key) is already a settable expression

set already works for cdr, and we can use cdr as above to create a
set-able tabref expression. We could almost just replace tabref with
cdr:tabloc, except that cdr:tabloc will create empty entries in the
table whenever a key is not present, which is not what we want outside
of a set.

To deal with that, I created a new macro called "ifwhere", which takes
two arguments. ifwhere evaluates to its first argument if it is inside a
where expression, and its second argument otherwise. Now we can replace
tabref with an ifwhere expression. This is done in src/extensions.bel,
which is a new bel file that is concatenated on to the end of bel.bel.