playground: add a stateful endpoint
This lets us use the background like a REPL, with persistent mutable
state.

This commit introduces a new endpoint, POST /stateful, which takes a
state parameter as an input (along with the expression to be evaluated)
and returns an updated state parameter as an output (along with the
result of evaluating the expression).