print spaces in strings as " ", not "\sp"
In other words, "hello world", not "hello\spworld"

Space is a special case. The character can be written as either like a
printable character (\ ) or like a non-printable character (\sp).
Non-printable characters are printed with their escape codes when
rendered in strings. Space was included in that group because it has an
escape code. But we don't want it to be, so we add a special case for it
in rendering.