def, mac, set now predefined, not "special forms"
def, mac, and set are in a unique class of Bel operators. (This class
includes bquote, comma, and comma-at, but that is not dealt with in this
commit.) These operators are not primitives, and they are not special
forms. They are defined in bel.bel, along with many other macros and
functions. This interpreter defines most of those macros and functions
simply by loading bel.bel. But def, mac, and set cannot be loaded that
way, because bel.bel _refers_ to def, mac, and set in the process of
defining them.

So while not being primitives or special forms, they still need to be
predefined before loading bel.bel. Previously we were predefining them
by grouping them with the special forms. But this is not great for two
reasons: (1) it's confusing for the set of "special forms" in this
implementation to be different from the set of special forms defined in
the Bel spec. (2) special forms do not go in global scope, which means
they cannot be looked up by user code. This has caused problems for me
when running the function "bel" -- Paul Graham's interpreter defined in
Bel user code. It expects to be able to look up names that are not
special forms. This has caused issues for me with backquoting in
particular.