uncomment test and record spec bug
This test was commented out because it was failing. After investigating
why, it turns out that the interpreter is behaving correctly and the
example in the spec is wrong, because (pr nil) somewhat surprisingly
prints no characters. This is in contrast with (prn nil) and (print
nil), which both print the letters n-i-l.