bug fix: don't eval after's 2nd arg in crash stack
The way after was previously written, when evaluating (after x y), if
the evaluation of x caused an error, the evaluation of y would proceed
_from the stack frame in which x crashed_, meaning among other things
that variables that y was expecting to be in scope were no longer in
scope.

After this commit, the evaluation of y now proceeds in the same stack
frame as the evaluation of after.

This change also throws away the crashing stack frames. We will probably
want to change that later.