cleaner representation of native operators
There are some operators which are neither primitives nor special forms
but are nonetheless implemented natively in the interpreter. The two
reasons for implementing operators this way are to make them faster or
to break circular definition chains in bel.bel. See this commit for more
on these: 513160292deb8e6af3b5783bd6bf0e5523be4142

These native operators are currently represented as special pairs called
OptimizedFunction. It is not visible from the Bel userland that these
operators are special. From Bel userland, they look like regular
functions and macros. But if you try to inspect them in Bel userland,
you "collapse" them -- they are transformed into the actual Bel code
they are masquerading as, and no longer use the optimized native
implementation. They do this because they have no good alternative on
inspection -- if the user were to modify the structure they saw, but the
function _didn't_ collapse, then the function's behavior would no longer
match the code that it was pretending to be.

This "collapse" behavior is undesirable both because it adds complexity
to the implementation and because it occasionally causes needless
slowdowns when we revert a function to its non-optimized version -- a
slowdown that can only be fixed by restarting the interpreter.

Relevant side note: the main time that collapse happens for me is when I
evaluate the function "bel" from bel.bel, which needs to inspect the
structure of functions in order to interpret them.

This commit introduces a new way to represent natively-implemented
operators: a new lit type, called "native". As an example, here is the
representation of the native implementation of "no":

    (lit native no)

There is nothing special going on behind the scenes here. This is a
regular list in both Bel userland and in the inner workings of the
interpreter. You can inspect it, modify it, and cons up equivalent lists
yourself.

The behind-the-scenes magic is introduced when a literal of this type is
used as an operator. The interpreter has a special case for operators of
the form (lit native x) -- it looks up a native implementation of x, and
uses it if it has one.

Furthermore, we also have a new virfn for native literals. Here it is:

  (vir native (f args)
    `(,(cdr (get (caddr f) natives)) ,@args))

This means, when using (lit native x) as an operator, look up x in the
"natives" key-value list, and run the code that you find there. This
virfn is our way of having optimized functions masquerade as regular Bel
functions. Our interpreter does not actually use this virfn, because it
already has special-case handling for native literals, but if it didn't
have the special-casing for them then these calls would still work
because they would go through the virfn and execute the equivalent bel
code.

In particular, having the virfn means that the "bel" function from
bel.bel can evaluate our native literals, despite not having access to
the native implementations. And it can do so without causing the
"collapse" behavior from before.

This commit only changes one native operator: the function "no". The
rest of the operators will be changed shortly.