convert a bunch of native operators to the new rep
See previous commit.

This commit converts a bunch of operators to the new (lit native x)
representation.

It also changes how we look up the native implementations. We now look
them up in a map rather than a list, so lookups are much faster. (This
didn't matter at the last commit when there was only one operator in the
list.)

The list of operators converted includes set, def, and mac, along with a
all the other "predefined" operators. This is significant because until
this commit those operators did not have their fallback code set. They
were masquerading to the Bel userland not as their code as defined in
bel.bel, but as the list (nil). This means that the "bel" function in
bel.bel could not interpret them, which in turn meant that load didn't
work, which in turn means that one of our tests fails. (This is our last
failing test, in fact). As of this commit, these operators now have
their virfn-based fallbacks set, which means the "bel" function can
interpret them.

See these commits for more info about these predefined operators:

    513160292deb8e6af3b5783bd6bf0e5523be4142
    def, mac, set now predefined, not "special forms"

    3168d86b50952d61cb5fb098f9f5e27b1a2707c8
    better (more correct) implementation of backquote

To set the fallbacks for these operators, I pasted their definitions
into extensions.bel, after having previously deleted them from
prelude.bel. They are now defined at the very end of extensions.bel, and
then immediately have their definitions moved into natives and their
global bindings replaced with (lit native ...)

I had to change the definition of set a bit to make the fallback work.
Rather than `(mac set..` as defined in bel.bel, I do
`(native set macro...`. This skips adding set to the global bindings,
and instead adds it directly to the native key-value list. This is
needed because the code that handles the rest of the fallbacks after
loading extensions.bel itself calls set, and the implementation of set
from bel.bel requires a 2-argument version of where, which is not yet
implemented in this interpreter.

Unfortunately, this commit also exposes a way in which the
(lit native x) approach is not spec-compliant: When a function f is
native-ized this way, (function f) will now return nil rather than clo.
This is exposed in the test:

    "(map function (list car append 'foo))" `is` "(prim clo nil)"

I worked around this by modifying function to recognize native literals
as well, and changing the test to expect the result (prim native nil)
here. Ideally I would like to implement all of the examples from
reference/bellanguage.txt exactly as written (minus the ones that were
written incorrectly), but I don't see a good way to do that here while
also optimizing these functions.

On seeing the above test failure, I considered a variation on my
previous approach: rather than collapsing OptimizedFunction pairs when
they are inspected, I could create a notion of read-only pairs, which
the bel-in-bel interpreter could read and interpret without issue, but
which would forbid users from modifying the facade of optimized
functions, and getting into a state in which these functions behave
differently from the code they are masquerading as. But this approach
cannot prevent that outcome. In fact, the existing "collapse" approach
already could not prevent that outcome. The reason is that the user does
not have to modify a function's code to modify its behavior. They can
also modify any function that said function calls. So to guarantee
correspondance between optimized functions and their code facades, if I
am using facades, I have to prevent any modifications to the facades,
prevent modifications to functions called by the facades, and prevent
modifications to the name bindings of any functions named by the
facades. This seems fairly restrictive, and quite plausibly worse than
the (lit native ...) approach that I am currently switching to.