convert more native operators to the new rep
See these previous commits:

    c1e3f68 convert a bunch of native operators to the new rep
    f2bd848 cleaner representation of native operators

These operators all have only partial native implementations: when the
inputs are numbers, we use a native implementation. For other inputs, we
fall back to the code defined in the spec. To implement that behavior,
this commit adds a bel function called "spec". (spec x) returns the code
in the spec for function x (only for functions that now have native
implementations).