simplify the process of loading interpreter state
We have a mechanism to serialize interpreter state to string and
deserialize it back to the state object. Previously, deserialization
required an additional step of modifying all of the serialized operators
that have native implementations, to replace them with an
invisible-to-bel-users, in-memory-only structure. Now there are no such
structures to insert, and we can simplify the process. This commit does
so.