playground: shrink /stateful state size by 99%
The serialization of the state created by bel.bel is about 40kb.
Compression brings that down to around 20kb. Before this commit,
interactions with the /stateful endpoint sent that 20kb back and forth
every time.

But we don't need to send that back and forth, since it is basically the
same every time. This commit does a simple text diff before compressing
the state, so we are now only passing back and forth diffs of the state
against the base state from bel.bel. For typical expected usages of the
POST /stateful endpoint, with at most a few variables set to relatively
small values, this will be at most a few hundred bytes, in contrast to
the previous ~20kb.

Sadly, this reduction in state size does not make the /stateful endpoint
any faster. It is still a few times slower than the root, non-stateful
endpoint. It looks like we are spending most of our time serializing
the state to string, before diffing and compressing it.