ci: update base docker image
This should fix the build. The previous image was built on the previous
stack snapshot. After I updated the stack snapshot, the CI build failed.

This version of the image shares a significant prefix with
playground/Dockerfile, so we get cache hits between the two.

Ideally I will eventually switch CI away from using a Docker image, in
favor of the GitHub action cache.