(attempt to) speed up repr with string builders
I implemented this because POST /stateful{-long} in the playground was
slow, and was spending most of its time serializing state to string,
which in turn means spending most of its time in repr.

repr is implemented in terms of String and does a lot of naive String
appends. It seemed plausible that this was asymptotically very slow, and
easy to fix by using a builder pattern instead.

This commit is my implementation of the builder pattern. It was easy
enough to implement, but yielded no apparent speed gains :/

I'm going to immediately revert this commit, because it is added
complexity without apparent benefit, but I am leaving it in the history
in case I want to revisit this in the future.