minimize compile time when only *.bel code changes
Previously, when I edited extensions.bel and re-compiled the
interpreter, most of the compilation time was spent re-building Eval.hs,
because Eval.hs held the embedded copy of the prelude *.bel files.

That re-compilation time is kind of a waste when I'm iterating on *.bel
files. Eval.hs doesn't really need to be the place where I hold the
embedded files. Moving them out saves a decent chunk of the compilation
time.

I also moved some more code out of that re-compilation path while I was
at it.