new representation for natively-implemented ops
I have done a lot of iterating on this representation, and I am still
not yet satisfied.

See previous work here:

    b317334 delete OptimizedFunction
    d069111 convert all remaining native ops to the new rep
    f43329e convert more native operators to the new rep
    c1e3f68 convert a bunch of native operators to the new rep
    f2bd848 cleaner representation of native operators

This is now the fourth representation I have implemented for these
operators. The first was OptimizedFunction -- an opaque-to-userland
representation. The second was (lit native _), and was the subject of
all of the above-named commits. The third never reached the commit log,
but was similar to the fourth, which is implemented in this commit. Here
it is, sketched as Bel code:

  (set hook (uvar))
  `(lit clo ((,hook . fn-name) ((nil) "This function is implemented natively in Chime.)) nil nil)

The new representation places a special pair as the first variable in a
closure's environment, whose value is the name of the Chime-internal
function (which is the same name as the corresponding Bel function). It
also places a short doc string in the environment for users to see. The
params and body of the closure are nil.

Here are some of my design criteria for a good representation:
- (required) a hook for Chime to swap in its native function at eval time
- should be obvious to users that Chime is not running Bel code
- should be spec compliant
  (not clear exactly what I mean here, but (lit native _) failed this by
  requiring one of the tests to change)
- the interpreter defined in bel.bel should be able to evaluate it

This last criterion, sadly, appears to be impossible. As far as I can
tell, the interpreter in bel.bel is not capable of hosting itself. Since
Chime is meant to be logically equivalent to that interpreter, this also
means that Chime cannot host it. This in turn means that even if I were
to throw out the native functions and use exactly the definitions
present in bel.bel, the interpreter defined in bel.bel would not be able
to evaluate (many of) them.

The bel.bel cannot host itself (as far as I can tell) is that enters an
infinite loop when trying to evaluate the macro "bquote": some of the
functions and macros transitively called by bquote are defined with
backquotes. bel.bel implements backquotes by re-writing them to
invocations of "bquote", which means the inner interpreter will recur on
evaluating bquote on the exact same arguments and environment, which is
an infinite loop.

Until I find a good way to solve this problem, I have given up on that
last design criterion. The current representation is satisfactory on the
remaining measures, so I am going with it.

I only discovered this self-hosting issue in bel.bel by pushing hard on
getting Chime to host bel.bel, which in turn meant pushing hard on this
problem of representing native functions. There was a lot of turnover
and I am still not satisfied, but I think it has been valuable.