fix a regression
The "spec" fallback mechanism was part of the (lit native _) thing that
I removed in the last commit. But I'd forgotten that I also use it as a
fallback for numeric functions. I'm adding it back to fix the numeric
functions.

This was caught by CI. I didn't catch it before pushing because the
failure was in one of the "slow" tests, which I don't run regularly.