implement load. all spec tests now pass!!! 🎉🎉🎉
The last remaining failing test was easy to implement with a native
function.

I had been hoping to use the version in bel.bel, but this proved to be
impossible, for reasons discussed at the end of the message in:

    3439e4d new representation for natively-implemented ops

But the native version is very simple and very fast, so I'll take it.

And I have now crossed a huge milestone!