move playground to its own package
The playground (the backend for the web-based repl, currently hosted at
https://bel-repl.com) should probably be in its own repo. It's here for
now because it is convenient.

The playground slows down build times of the interpreter. It has to be
built the first time you build the interpreter (because of some cabal
limitation I don't understand). It has some large dependencies, plus
building any haskell code is slow and so I want users who build from
source to have to do as little of it as possible.

It also isn't building for me on Windows. (The specific error is from
regex-posix, a dependency that the playground has but the interpreter
doesn't have: "regex.h: No such file or directory".) This is stopping
the interpreter from building on Windows as well, which is bad, but easy
to fix (I think) by pulling the playground into its own package.