playground: implement sharing feature
On bel-repl.com you can now click a "share" button to get a sharable
link to the code you have written in the editor. This is a really nice
feature that I copied from the Go playground (play.golang.org). I have
used the Go playground version many times.

Ignoring hosting issues, it's a dead-simple feature: hash the text,
store the text in a database keyed by its hash, and return the hash to
the frontend.

The database I've chosen is Google Cloud Storage. I created a GCS bucket
called chime-snippets. The bucket is publicly readable, but only
writable by me and a service account I created. A lot of the code in
this commit is just managing secret keys for that account. Hopefully I
didn't mess it up!

I needed to do some work with ByteStrings, Text and base64 encoding in
this commit, and found that I didn't like some of the existing patterns
I was using, so I went back and changed some of those (eg the changes in
compressState and decompressState). I changed "LazyText" to "LT" and
swapped out the base64 library I was using.